+++ vllm/multimodal/registry.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 from collections.abc import Mapping
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Any, Generic, Protocol, TypeVar
+from typing import TYPE_CHECKING, Generic, Protocol, TypeVar
 
 import torch.nn as nn
 
@@ -358,20 +358,3 @@ class MultiModalRegistry:
 
         first_modality = next(iter(max_tokens))
         return max_tokens[first_modality]
-
-
-class ExtensionManager:
-    def __init__(self) -> None:
-        self.name2class: dict[str, type] = {}
-
-    def register(self, name: str):
-        def wrap(cls_to_register):
-            self.name2class[name] = cls_to_register
-            return cls_to_register
-
-        return wrap
-
-    def load(self, cls_name: str, *args, **kwargs) -> Any:
-        cls = self.name2class.get(cls_name)
-        assert cls is not None, f"Extension class {cls_name} not found"
-        return cls(*args, **kwargs)
